requirements.txt
fastapi
uvicorn
streamlit
langchain
langgraph
pydantic
numpy
pypdf
openpyxl
pandas
xlrd
pymilvus
python-dotenv
httpx
requests
websocket-client